Extract more color picker related logic to `ColorHoverParticipant`
